Add some additional functions to QMF schema APIs.
This patch adds some additional functions to schema related QMF APIs.
When looking through a schema, the APIs provided allow you to get a
count of objects and then to access them by index.  It seems more
natural to look up these items by name, so various getter functions have
been added to allow for this.

The one exception is hasPackage(), which is simply to be able to check
to see if an Agent has a particular package.  The added get functions
for other objects provide for this use case (among others) since they
implement operator bool.

Signed-off-by: Russell Bryant <rbryant@redhat.com>